+++ btr-web/btr-common-components/cypress.config.js
@@ -2,6 +2,7 @@ import { defineConfig } from 'cypress'
 
 export default defineConfig({
   e2e: {
-    baseUrl: 'http://localhost:3050'
+    baseUrl: 'http://localhost:3050',
+    supportFile: false
   }
 })
