+++ btr-web/btr-common-components/pages/index.vue
@@ -2,7 +2,7 @@
   <div class="m-10">
     <h1>Links ot examples:</h1>
     <br>
-    <NuxtLink to="/examples/forms">
+    <NuxtLink to="/examples/form">
       Forms example
     </NuxtLink>
   </div>
