+++ btr-web/btr-common-components/vitest.config.js
@@ -1,0 +1,7 @@
+import { defineVitestConfig } from 'nuxt-vitest/config'
+
+export default defineVitestConfig({
+  test: {
+    environment: 'nuxt'
+  }
+})
