+++ btr-web/btr-layouts/package.json
@@ -18,26 +18,6 @@
     "test": "vitest",
     "test:cov": "vitest run --coverage"
   },
-  "devDependencies": {
-    "@nuxt/devtools": "latest",
-    "@nuxt/test-utils": "^3.7.4",
-    "@nuxt/ui": "^2.9.0",
-    "@nuxtjs/eslint-config-typescript": "^12.1.0",
-    "@nuxtjs/eslint-module": "^4.1.0",
-    "@nuxtjs/stylelint-module": "^5.1.0",
-    "@pinia/nuxt": "^0.4.11",
-    "cypress": "^13.3.0",
-    "eslint": "^8.50.0",
-    "nuxt": "^3.7.4",
-    "nuxt-vitest": "^0.10.5",
-    "postcss-html": "^1.5.0",
-    "sass": "^1.68.0",
-    "sass-loader": "^13.3.2",
-    "stylelint": "^15.10.3",
-    "stylelint-config-recommended-vue": "^1.5.0",
-    "typescript": "^5.2.2",
-    "vitest": "^0.33.0"
-  },
   "simple-git-hooks": {
     "pre-commit": "npx lint-staged"
   },
