+++ btr-web/btr-main-app/.npmrc
@@ -1,0 +1,2 @@
+shamefully-hoist=true
+shared-workspace-lockfile=false
